Created Checkout container and logic
In Checkout.tsx container displays selected orders with their quantity, price, price for delivery and total price.

By clicking on the remove button on the OrderCard.tsx, selected order will be removed after confirming.
By clicking Order button, it will be posted to the firebase, Order list will be cleaned as a total price and user will be navigated to the main page.